Move src/core to src/utils
