+++ src/content-scripts/utils/Highlighter.ts
