+++ src/content-scripts/utils/index.ts
@@ -1,0 +1,3 @@
+export { default as Highlighter } from './Highlighter';
+export { default as TextSelector } from './TextSelector';
+export { default as DragTracker } from './DragTracker';
